+++ verl/trainer/main_ppo.py
@@ -11,45 +11,26 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""
-Note that we don't combine the main with ray_trainer as ray_trainer is used by other mpain.
-"""
-
+import logging
 import os
-import socket
+from pprint import pprint
 
 import hydra
 import ray
-from omegaconf import OmegaConf
+from omegaconf import DictConfig, OmegaConf
 
-from verl.experimental.reward_loop import migrate_legacy_reward_impl
 from verl.trainer.constants_ppo import get_ppo_ray_runtime_env
-from verl.trainer.distillation import is_distillation_enabled
-from verl.trainer.ppo.ray_trainer import RayPPOTrainer
 from verl.trainer.ppo.utils import need_critic, need_reference_policy
 from verl.utils.config import validate_config
 from verl.utils.device import auto_set_device, is_cuda_available
-from verl.utils.import_utils import deprecated
+from verl.utils.import_utils import load_class_from_fqn
 
-
-@deprecated(
-    "main_ppo.py is deprecated, and wil be replaced by main_ppo_sync.py in v0.8.0, please use main_ppo_sync.py instead."
-)
-@hydra.main(config_path="config", config_name="ppo_trainer", version_base=None)
-def main(config):
-    """Main entry point for PPO training with Hydra configuration management.
-
-    Args:
-        config: Hydra configuration dictionary containing training parameters.
-    """
-    # Automatically set `config.trainer.device = npu` when running on Ascend NPU.
-    auto_set_device(config)
-    config = migrate_legacy_reward_impl(config)
-    run_ppo(config)
+logger = logging.getLogger(__name__)
+logger.setLevel(os.getenv("VERL_LOGGING_LEVEL", "INFO"))
 
 
 # Define a function to run the PPO-like training process
-def run_ppo(config, task_runner_class=None) -> None:
+def run_ppo(config, task_runner_class) -> None:
     """Initialize Ray cluster and run distributed PPO training process.
 
     Args:
@@ -79,9 +60,6 @@ def run_ppo(config, task_runner_class=None) -> None:
         print(f"ray init kwargs: {ray_init_kwargs}")
         ray.init(**OmegaConf.to_container(ray_init_kwargs))
 
-    if task_runner_class is None:
-        task_runner_class = ray.remote(num_cpus=1)(TaskRunner)  # please make sure main_task is not scheduled on head
-
     # Create a remote instance of the TaskRunner class, and
     # Execute the `run` method of the TaskRunner instance remotely and wait for it to complete
     if (
@@ -108,272 +86,88 @@ def run_ppo(config, task_runner_class=None) -> None:
         ray.timeline(filename=timeline_json_file)
 
 
-class TaskRunner:
-    """Ray remote class for executing distributed PPO training tasks.
-
-    This class encapsulates the main training logic and runs as a Ray remote actor
-    to enable distributed execution across multiple nodes and GPUs.
-
-    Attributes:
-        role_worker_mapping: Dictionary mapping Role enums to Ray remote worker classes
-        mapping: Dictionary mapping Role enums to resource pool IDs for GPU allocation
-    """
+@ray.remote
+class TaskRunnerV1:
+    """V1 TaskRunner for PPO training."""
 
     def __init__(self):
-        self.role_worker_mapping = {}
-        self.mapping = {}
+        self.config = None
+        self.trainer = None
+        self.agent_loop_manager = None
 
-    def add_actor_rollout_worker(self, config):
-        """Add actor rollout worker using the unified model engine implementation."""
-        from verl.single_controller.ray import RayWorkerGroup
-        from verl.trainer.ppo.ray_trainer import Role
-        from verl.workers.engine_workers import ActorRolloutRefWorker
+    def init_agent_loop_manager(self):
+        """Initialize the agent loop manager to generate sequences.
 
-        actor_rollout_cls = ActorRolloutRefWorker
-        ray_worker_group_cls = RayWorkerGroup
-
-        lora_rank = config.actor_rollout_ref.model.get("lora", {}).get("rank", 0)
-        if lora_rank <= 0:
-            lora_rank = config.actor_rollout_ref.model.get("lora_rank", 0)
-        ref_in_actor = lora_rank > 0 or config.actor_rollout_ref.model.get("lora_adapter_path") is not None
-        # Ref policy is fused into ActorRolloutRefWorker unless LoRA is used with a dedicated ref model.
-        if need_reference_policy(config) and not ref_in_actor:
-            role = Role.ActorRolloutRef
-        else:
-            role = Role.ActorRollout
-        self.role_worker_mapping[role] = ray.remote(actor_rollout_cls)
-        self.mapping[role] = "global_pool"
-        return actor_rollout_cls, ray_worker_group_cls
-
-    def add_critic_worker(self, config):
-        """Add critic worker to role mapping using the unified model engine implementation."""
-        from verl.trainer.ppo.ray_trainer import Role
-        from verl.workers.engine_workers import TrainingWorker
-
-        # The model-engine TrainingWorker handles all critic backends (fsdp/fsdp2/megatron/...)
-        # internally based on ``config.critic.strategy``.
-        self.role_worker_mapping[Role.Critic] = ray.remote(TrainingWorker)
-        self.mapping[Role.Critic] = "global_pool"
-
-    def init_resource_pool_mgr(self, config):
-        """Initialize resource pool manager."""
-
-        global_pool_id = "global_pool"
-        resource_pool_spec = {
-            global_pool_id: [config.trainer.n_gpus_per_node] * config.trainer.nnodes,
-        }
-
-        if config.reward.reward_model.enable_resource_pool:
-            if config.reward.reward_model.n_gpus_per_node <= 0:
-                raise ValueError("config.reward.reward_model.n_gpus_per_node must be greater than 0")
-            if config.reward.reward_model.nnodes <= 0:
-                raise ValueError("config.reward.reward_model.nnodes must be greater than 0")
-
-            reward_pool = [config.reward.reward_model.n_gpus_per_node] * config.reward.reward_model.nnodes
-            resource_pool_spec["reward_pool"] = reward_pool
-        else:
-            config.reward.reward_model.nnodes = config.trainer.nnodes
-            config.reward.reward_model.n_gpus_per_node = config.trainer.n_gpus_per_node
-
-        distillation_config = config.get("distillation")
-        if is_distillation_enabled(distillation_config):
-            if distillation_config.n_gpus_per_node <= 0:
-                raise ValueError("config.distillation.n_gpus_per_node must be greater than 0")
-            if distillation_config.nnodes <= 0:
-                raise ValueError("config.distillation.nnodes must be greater than 0")
-
-            teacher_pool = [distillation_config.n_gpus_per_node] * distillation_config.nnodes
-            resource_pool_spec["teacher_pool"] = teacher_pool
-
-        from verl.trainer.ppo.ray_trainer import ResourcePoolManager
-
-        resource_pool_manager = ResourcePoolManager(resource_pool_spec=resource_pool_spec, mapping=self.mapping)
-        return resource_pool_manager
-
-    def add_reward_model_resource_pool(self, config):
-        """Add reward model worker if enabled."""
-        from verl.trainer.ppo.ray_trainer import Role
-
-        if config.reward.reward_model.enable:
-            # we do not use reward model workers, so we only register reward model in resource pool
-            # without continue to register reward model worker in role mapping
-            if config.reward.reward_model.enable_resource_pool:
-                self.mapping[Role.RewardModel] = "reward_pool"
-            else:
-                self.mapping[Role.RewardModel] = "global_pool"
-
-    def add_teacher_model_resource_pool(self, config):
-        """Add teacher model worker if enabled."""
-        from verl.trainer.ppo.ray_trainer import Role
-
-        if is_distillation_enabled(config.get("distillation")):
-            # we do not use teacher model workers, so we only register teacher model in resource pool
-            # without registering a teacher model worker in role-worker mapping
-            self.mapping[Role.TeacherModel] = "teacher_pool"
-
-    def add_ref_policy_worker(self, config, ref_policy_cls):
-        """Ref policy is fused into ActorRolloutRefWorker in the unified model engine.
-
-        Kept for backward compatibility with subclasses that still invoke it; the method
-        is now a no-op because the reference policy lives on the same worker group as
-        the actor/rollout.
+        NOTE: User can customize their own agent loop manager, the only requirement is:
+        1. implement `generate_sequences` method
+        2. put agent loop outputs into TransferQueue
         """
-        return
+        from verl.trainer.ppo.v1 import AgentLoopManagerTQ
 
-    def run(self, config):
-        """Execute the main PPO training workflow.
+        manager_class_fqn = self.config.actor_rollout_ref.rollout.get("agent", {}).get("agent_loop_manager_class")
+        if manager_class_fqn:
+            agent_loop_manager_cls = load_class_from_fqn(manager_class_fqn, "AgentLoopManager")
+        else:
+            agent_loop_manager_cls = AgentLoopManagerTQ
 
-        This method sets up the distributed training environment, initializes
-        workers, datasets, and reward functions, then starts the training process.
+        self.agent_loop_manager = agent_loop_manager_cls.create(
+            config=self.config,
+            llm_client=self.trainer.get_llm_client(),
+            teacher_client=self.trainer.get_teacher_client(),
+            reward_loop_worker_handles=self.trainer.get_reward_handles(),
+        )
 
-        Args:
-            config: Training configuration object containing all parameters needed
-                   for setting up and running the PPO training process.
-        """
-        # Print the initial configuration. `resolve=True` will evaluate symbolic values.
-        from pprint import pprint
+    def run(self, config: DictConfig):
+        """Run the PPO training process."""
+        import transfer_queue as tq
 
-        from omegaconf import OmegaConf
+        from verl.trainer.ppo.v1 import get_trainer_cls
 
-        from verl.utils.fs import copy_to_local
+        trainer_cls = get_trainer_cls(config.trainer.v1.trainer_mode)
 
-        print(f"TaskRunner hostname: {socket.gethostname()}, PID: {os.getpid()}")
+        config.transfer_queue.enable = True
         pprint(OmegaConf.to_container(config, resolve=True))
         OmegaConf.resolve(config)
+        self.config = config
 
-        actor_rollout_cls, ray_worker_group_cls = self.add_actor_rollout_worker(config)
-        self.add_critic_worker(config)
-
-        self.add_reward_model_resource_pool(config)
-
-        self.add_teacher_model_resource_pool(config)
-
-        # Add a reference policy worker if KL loss or KL reward is used.
-        self.add_ref_policy_worker(config, actor_rollout_cls)
-
-        # validate config
-        validate_config(
-            config=config,
-            use_reference_policy=need_reference_policy(config),
-            use_critic=need_critic(config),
-        )
-
-        # Download the checkpoint from HDFS to the local machine.
-        # `use_shm` determines whether to use shared memory, which could lead to faster model loading if turned on
-        local_path = copy_to_local(
-            config.actor_rollout_ref.model.path, use_shm=config.actor_rollout_ref.model.get("use_shm", False)
-        )
-
-        # Instantiate the tokenizer and processor.
-        from verl.utils import hf_processor, hf_tokenizer
-
-        trust_remote_code = config.data.get("trust_remote_code", False)
-        tokenizer = hf_tokenizer(local_path, trust_remote_code=trust_remote_code)
-        # Used for multimodal LLM, could be None
-        processor = hf_processor(local_path, trust_remote_code=trust_remote_code, use_fast=True)
-
-        resource_pool_manager = self.init_resource_pool_mgr(config)
-
-        from verl.utils.dataset.rl_dataset import collate_fn
-
-        # Create training and validation datasets.
-        train_dataset = create_rl_dataset(
-            config.data.train_files,
-            config.data,
-            tokenizer,
-            processor,
-            is_train=True,
-            max_samples=config.data.get("train_max_samples", -1),
-        )
-        val_dataset = create_rl_dataset(
-            config.data.val_files,
-            config.data,
-            tokenizer,
-            processor,
-            is_train=False,
-            max_samples=config.data.get("val_max_samples", -1),
-        )
-        train_sampler = create_rl_sampler(config.data, train_dataset)
-
-        # Initialize the PPO trainer.
-        trainer = RayPPOTrainer(
-            config=config,
-            tokenizer=tokenizer,
-            processor=processor,
-            role_worker_mapping=self.role_worker_mapping,
-            resource_pool_manager=resource_pool_manager,
-            ray_worker_group_cls=ray_worker_group_cls,
-            train_dataset=train_dataset,
-            val_dataset=val_dataset,
-            collate_fn=collate_fn,
-            train_sampler=train_sampler,
-        )
-        # Initialize the workers of the trainer.
-        trainer.init_workers()
-
-        # Start the training process.
-        trainer.fit()
+        # initialize transfer queue
+        tq.init(config.transfer_queue)
+        try:
+            self.trainer = trainer_cls(config=config)
+            self.trainer.init()
+            self.init_agent_loop_manager()
+            self.trainer.fit(self.agent_loop_manager)
+        finally:
+            tq.close()
 
 
-def create_rl_dataset(data_paths, data_config, tokenizer, processor, is_train=True, max_samples: int = -1):
-    """Create a dataset.
+@hydra.main(config_path="config", config_name="ppo_trainer", version_base=None)
+def main(config):
+    """Main entry point for PPO training with Hydra configuration management.
 
-    Arguments:
-        data_paths: List of paths to data files.
-        data_config: The data config.
-        tokenizer (Tokenizer): The tokenizer.
-        processor (Processor): The processor.
-
-    Returns:
-        dataset (Dataset): The dataset.
+    Args:
+        config: Hydra configuration dictionary containing training parameters.
     """
+    # Automatically set `config.trainer.device = npu` when running on Ascend NPU.
+    auto_set_device(config)
 
-    from verl.utils.dataset.rl_dataset import get_dataset_class
-
-    # Get the dataset class
-    dataset_cls = get_dataset_class(data_config)
-
-    # Instantiate the dataset using the determined dataset class
-    dataset = dataset_cls(
-        data_files=data_paths,
-        tokenizer=tokenizer,
-        processor=processor,
-        config=data_config,
-        max_samples=max_samples,
+    # validate config
+    validate_config(
+        config=config,
+        use_reference_policy=need_reference_policy(config),
+        use_critic=need_critic(config),
     )
 
-    return dataset
-
-
-def create_rl_sampler(data_config, dataset):
-    """Create a sampler for the dataset.
-
-    Arguments:
-        data_config: The data config.
-        dataset (Dataset): The dataset.
-
-    Returns:
-        sampler (Sampler): The sampler.
-    """
-    import torch
-    from torch.utils.data import SequentialSampler
-
-    # torch.utils.data.RandomSampler could not recover properly
-    from torchdata.stateful_dataloader.sampler import RandomSampler
-
-    # Use a sampler to facilitate checkpoint resumption.
-    # If shuffling is enabled in the data configuration, create a random sampler.
-    if data_config.shuffle:
-        train_dataloader_generator = torch.Generator()
-        seed = data_config.get("seed")
-        if seed is not None:
-            train_dataloader_generator.manual_seed(seed)
-        sampler = RandomSampler(data_source=dataset, generator=train_dataloader_generator)
+    if config.trainer.use_v1:
+        run_ppo(config, task_runner_class=TaskRunnerV1)
     else:
-        # If shuffling is disabled, use a sequential sampler to iterate through the dataset in order.
-        sampler = SequentialSampler(data_source=dataset)
+        from verl.trainer.main_ppo_v0 import TaskRunner
 
-    return sampler
+        logger.warning(
+            "Legacy trainer `main_ppo_v0.py` is deprecated, and wil be removed in v0.9.0."
+            "Please set `trainer.use_v1=True` in config to use V1 trainer."
+        )
+        run_ppo(config, task_runner_class=TaskRunner)
 
 
 if __name__ == "__main__":
