--- verl/trainer/main_ppo_sync.py
@@ -1,1688 +1,0 @@
-# Copyright 2024 Bytedance Ltd. and/or its affiliates
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-"""
-Synchronous PPO trainer with colocated actor and rollout.
-
-Differs from original PPO trainer in main_ppo.py:
-1. Use TransferQueue to zero-padding and zero-copy data transfer.
-2. Use ReplayBuffer to sample data from TransferQueue.
-3. Support different `n` sampling for each prompt.
-4. Support multiple outputs for each agent loop.
-"""
-
-import asyncio
-import json
-import logging
-import math
-import os
-import uuid
-from collections import defaultdict
-from concurrent.futures import ThreadPoolExecutor
-from functools import partial
-from pprint import pprint
-from typing import Any
-
-import hydra
-import numpy as np
-import ray
-import torch
-import transfer_queue as tq
-from omegaconf import DictConfig, OmegaConf, open_dict
-from tensordict import NonTensorData, NonTensorStack, TensorDict
-from torchdata.stateful_dataloader import StatefulDataLoader
-from tqdm import tqdm
-from transfer_queue import KVBatchMeta
-
-from verl.checkpoint_engine import CheckpointEngineManager
-from verl.experimental.agent_loop import (
-    AgentLoopManager,
-    AgentLoopOutput,
-    AgentLoopWorker,
-    get_trajectory_info,
-)
-from verl.experimental.reward_loop import RewardLoopManager
-from verl.experimental.teacher_loop import MultiTeacherModelManager
-from verl.protocol import DataProto, DataProtoFuture
-from verl.single_controller.ray import (
-    RayClassWithInitArgs,
-    RayWorkerGroup,
-    ResourcePoolManager,
-    create_colocated_worker_cls,
-)
-from verl.trainer.distillation import is_distillation_enabled
-from verl.trainer.main_ppo import create_rl_dataset, create_rl_sampler, run_ppo
-from verl.trainer.ppo import core_algos
-from verl.trainer.ppo.core_algos import agg_loss
-from verl.trainer.ppo.metric_utils import (
-    compute_data_metrics,
-    compute_throughout_metrics,
-    compute_timing_metrics,
-    compute_variance_proxy_metrics,
-    process_validation_metrics,
-)
-from verl.trainer.ppo.padding_utils import upsample_batch_to_divisible_size
-from verl.trainer.ppo.ray_trainer import apply_kl_penalty, compute_advantage, compute_spec_decode_metrics
-from verl.trainer.ppo.rollout_corr_helper import compute_rollout_correction_and_add_to_batch
-from verl.trainer.ppo.utils import Role, WorkerType, need_critic, need_reference_policy, need_teacher_policy
-from verl.trainer.ppo.v1.replay_buffer import ReplayBuffer
-from verl.utils import hf_processor, hf_tokenizer
-from verl.utils import tensordict_utils as tu
-from verl.utils.checkpoint.checkpoint_manager import find_latest_ckpt_path
-from verl.utils.config import omega_conf_to_dataclass, validate_config
-from verl.utils.dataset.rl_dataset import collate_fn
-from verl.utils.debug import marked_timer
-from verl.utils.debug.metrics import calculate_debug_metrics
-from verl.utils.device import auto_set_device
-from verl.utils.fs import copy_to_local
-from verl.utils.import_utils import load_class_from_fqn
-from verl.utils.metric import reduce_metrics
-from verl.utils.py_functional import rename_dict
-from verl.utils.ray_utils import auto_await
-from verl.utils.seqlen_balancing import calculate_workload, get_seqlen_balanced_partitions, log_seqlen_unbalance
-from verl.utils.tensordict_utils import list_of_dict_to_tensordict
-from verl.utils.tracking import Tracking, ValidationGenerationsLogger
-from verl.workers.config import CriticConfig, DistillationConfig
-from verl.workers.engine_workers import ActorRolloutRefWorker, TrainingWorker, TrainingWorkerConfig
-from verl.workers.rollout.llm_server import LLMServerManager
-from verl.workers.utils.losses import value_loss
-from verl.workers.utils.padding import response_from_nested, response_to_nested
-
-
-def apply_greedy_sampling_params(params: dict[str, Any]) -> None:
-    params["top_p"] = 1.0
-    params["top_k"] = -1
-    params["temperature"] = 0
-
-
-logger = logging.getLogger(__name__)
-logger.setLevel(os.getenv("VERL_LOGGING_LEVEL", "INFO"))
-
-
-# ======================================= USER SECTION BEGIN =======================================
-
-
-def compute_advantage_for_multi_trajectories(
-    data: DataProto,
-    batch_keys: list[str],
-    adv_estimator,
-    gamma: float = 1.0,
-    lam: float = 1.0,
-    num_repeat: int = 1,
-    norm_adv_by_std_in_grpo: bool = True,
-    config: Any = None,
-) -> DataProto:
-    """Compute GRPO advantages from each session's final output. For non-GRPO
-    estimators, such as GAE, are delegated to the original compute_advantage() unchanged.
-
-    For GRPO, only the final output in each ``{uid}_{session_id}`` group participates
-    in advantage computation, and the result is broadcast to the other outputs in
-    the same session. Sessions whose AgentLoop returns ``None`` simply do not appear
-    in ``batch_keys``. Non-GRPO estimators, such as GAE, are delegated to the
-    original ``compute_advantage()`` unchanged.
-    """
-    if adv_estimator != core_algos.AdvantageEstimator.GRPO:
-        return compute_advantage(
-            data,
-            adv_estimator=adv_estimator,
-            gamma=gamma,
-            lam=lam,
-            num_repeat=num_repeat,
-            norm_adv_by_std_in_grpo=norm_adv_by_std_in_grpo,
-            config=config,
-        )
-
-    # final session of each agent loop: {uid}_{session_id} => (index, row_index)
-    final_sessions: dict[str, tuple[int, int]] = {}
-    row_session_keys = []
-    for i, key in enumerate(batch_keys):
-        fields = key.rsplit("_", 2)
-        assert len(fields) == 3, f"Unexpected key format: {key}"
-        uid, session_id, index = fields[0], fields[1], int(fields[2])
-        session_key = f"{uid}_{session_id}"
-        row_session_keys.append(session_key)
-        if session_key not in final_sessions or final_sessions[session_key][0] < index:
-            final_sessions[session_key] = (index, i)
-
-    # final session indices in batch data
-    final_indices = []
-    session_key_to_local_index = {}
-    for session_key, (_, row_index) in final_sessions.items():
-        final_indices.append(row_index)
-        session_key_to_local_index[session_key] = len(final_indices) - 1
-    row_to_local_index = [session_key_to_local_index[session_key] for session_key in row_session_keys]
-
-    # select final sessions from batch data for group relative advantage computation
-    final_data = compute_advantage(
-        data.select_idxs(final_indices),
-        adv_estimator=adv_estimator,
-        gamma=gamma,
-        lam=lam,
-        num_repeat=num_repeat,
-        norm_adv_by_std_in_grpo=norm_adv_by_std_in_grpo,
-        config=config,
-    )
-    first_nnz_indices = final_data.batch["response_mask"].argmax(dim=1)
-    final_scores = final_data.batch["advantages"][torch.arange(len(final_data)), first_nnz_indices]
-
-    # scatter final scores to all rows in batch data
-    scores = final_scores[row_to_local_index]
-    scores = scores.unsqueeze(-1) * data.batch["response_mask"]
-
-    data.batch["advantages"] = scores
-    data.batch["returns"] = scores
-    return data
-
-
-@ray.remote
-class AgentLoopWorkerTQ(AgentLoopWorker):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        tq.init()
-        self.background_tasks = set()
-
-    async def generate_sequences(self, batch: TensorDict) -> None:
-        """Spawn agent loop for each sample in the batch without waiting for the results."""
-        validate = batch["validate"] if "validate" in batch else False
-        batch.pop("validate", None)
-        config = self.config.actor_rollout_ref.rollout
-        sampling_params = dict(
-            temperature=config.temperature,
-            top_p=config.top_p,
-            top_k=config.top_k,
-            repetition_penalty=1.0,
-            logprobs=config.calculate_log_probs,
-        )
-
-        # override sampling params for validation
-        if validate:
-            sampling_params["top_p"] = config.val_kwargs.top_p
-            sampling_params["top_k"] = config.val_kwargs.top_k
-            sampling_params["temperature"] = config.val_kwargs.temperature
-
-        # by default, we assume it's a single turn agent
-        if "agent_name" not in batch:
-            default_agent_loop = config.agent.default_agent_loop
-            batch["agent_name"] = NonTensorData(default_agent_loop)
-
-        trajectory_info = await get_trajectory_info(batch["global_steps"], batch["index"], validate)
-
-        # create background tasks for each sample in the batch
-        for i in range(len(batch)):
-            # TODO(wuxibin): add trace support
-            trace_this_sample = False
-            prompt = {}
-            for k, v in batch.items():
-                if isinstance(v, torch.Tensor):
-                    prompt[k] = v[i]
-                elif isinstance(v, NonTensorStack):
-                    prompt[k] = v[i].data
-                elif isinstance(v, NonTensorData):
-                    prompt[k] = v.data
-                else:
-                    logger.exception(f"Unsupported type {type(v)} for key {k}")
-
-            # “fire-and-forget” background tasks
-            task = asyncio.create_task(
-                self._run_prompt(prompt, sampling_params, trajectory=trajectory_info[i], trace=trace_this_sample)
-            )
-            self.background_tasks.add(task)
-            task.add_done_callback(self.background_tasks.discard)
-
-    async def _run_prompt(self, prompt: dict, sampling_params: dict, trajectory: dict, trace: bool = False) -> None:
-        """Spawn multiple agent loops in parallel according to rollout.n or rollout.val_kwargs.n."""
-        uid, partition_id = prompt["uid"], "train" if not trajectory["validate"] else "val"
-        await tq.async_kv_put(key=uid, partition_id=partition_id, tag={"status": "running"})
-        try:
-            # NOTE: user can dynamically adjust n for each sample here, e.g according to task difficulty.
-            config = self.config.actor_rollout_ref.rollout
-            n = prompt.pop("__rollout_n__", config.n if not trajectory["validate"] else config.val_kwargs.n)
-            do_sample = prompt.pop("__do_sample__", True)
-
-            run_sampling_params = dict(sampling_params)
-            if not trajectory["validate"] and not do_sample:
-                apply_greedy_sampling_params(run_sampling_params)
-
-            tasks = []
-            for i in range(n):
-                task = asyncio.create_task(
-                    self._run_agent_loop(
-                        run_sampling_params, trajectory=trajectory, trace=trace, session_id=i, **prompt
-                    )
-                )
-                tasks.append(task)
-            await asyncio.gather(*tasks)
-            await tq.async_kv_put(key=uid, partition_id=partition_id, tag={"status": "finished"})
-        except Exception as e:
-            logger.exception(f"Error in _run_prompt: {e}")
-            await tq.async_kv_put(key=uid, partition_id=partition_id, tag={"status": "failure"})
-
-    async def _agent_loop_postprocess(
-        self, output: AgentLoopOutput | list[AgentLoopOutput], validate, **kwargs
-    ) -> None:
-        """Put agent loop outputs into TransferQueue."""
-        uid, session_id = kwargs["uid"], kwargs["session_id"]
-        outputs = output if isinstance(output, list) else [output]
-        if not outputs:
-            logger.warning(f"Empty output for prompt {uid}_{session_id}")
-            return
-
-        await self._compute_score(outputs, kwargs=kwargs)
-
-        final_output = outputs[-1]
-        # TODO: Support output:list[AgentLoopOutput]
-        await self._compute_teacher_logprobs(
-            final_output,
-            prompt_ids=final_output.prompt_ids,
-            response_ids=final_output.response_ids,
-            validate=validate,
-            sample_kwargs=kwargs,
-        )
-
-        if final_output.reward_score is not None:
-            for output in outputs[:-1]:
-                output.reward_score = final_output.reward_score
-                output.extra_fields["reward_extra_info"] = final_output.extra_fields["reward_extra_info"]
-
-        # NOTE: agent loop may has multiple outputs, put each output into TransferQueue.
-        # key format: {uid}_{session_id}_{index}
-        # - uid: raw prompt uid from dataset
-        # - session_id: session id for rollout.n sampling
-        # - index: index of agent loop output
-        keys, fields, tags = [], [], []
-        for i, output in enumerate(outputs):
-            prompts = torch.tensor(output.prompt_ids, dtype=torch.int64)
-            responses = torch.tensor(output.response_ids, dtype=torch.int64)
-            input_ids = torch.cat([prompts, responses], dim=0)
-            attention_mask = torch.ones_like(input_ids, dtype=torch.int64)
-            multi_modal_inputs = self._compute_multi_modal_inputs(output, input_ids)
-            position_ids = self._compute_position_ids(
-                input_ids.unsqueeze(0), attention_mask.unsqueeze(0), multi_modal_inputs
-            ).squeeze(0)
-
-            keys.append(f"{uid}_{session_id}_{i}")
-            field = output.as_dict()
-            field.update(kwargs)
-            # do not store raw image/video
-            field.pop("multi_modal_data", None)
-            # TODO: uniform response_mask and loss_mask
-            field["loss_mask"] = field["response_mask"]
-            field["input_ids"] = input_ids
-            field["position_ids"] = position_ids
-            field["multi_modal_inputs"] = multi_modal_inputs
-            fields.append(field)
-            prompt_len, response_len = field["prompts"].size(0), field["responses"].size(0)
-            tags.append(
-                {
-                    "status": "success",
-                    "prompt_len": prompt_len,
-                    "response_len": response_len,
-                    "seq_len": prompt_len + response_len,
-                    # These tags are used for off-policy staleness control, if a trajectory
-                    # spans too many global steps, we need to filter it out.
-                    # global_steps: which global steps this sample is from dataloader
-                    "global_steps": kwargs["global_steps"],
-                    # min_global_steps: start generation model weights version of this trajectory
-                    "min_global_steps": field["extra_fields"].get("min_global_steps"),
-                    # max_global_steps: end generation model weights version of this trajectory
-                    "max_global_steps": field["extra_fields"].get("max_global_steps"),
-                }
-            )
-
-        await tq.async_kv_batch_put(
-            keys=keys,
-            fields=list_of_dict_to_tensordict(fields),
-            tags=tags,
-            partition_id="train" if not validate else "val",
-        )
-
-
-class AgentLoopManagerTQ(AgentLoopManager):
-    def __init__(self, *args, **kwargs):
-        self.agent_loop_workers_class = AgentLoopWorkerTQ
-        super().__init__(*args, **kwargs)
-
-    @classmethod
-    @auto_await
-    async def create(cls, *args, **kwargs):
-        """Create agent loop manager."""
-        instance = cls(*args, **kwargs)
-        await instance._init_agent_loop_workers()
-        return instance
-
-    def generate_sequences(self, prompts: TensorDict) -> None:
-        """
-        Dispatch input batch to agent loop workers without blocking. Workers should put agent loop outputs
-        into TransferQueue once an agent loop finished.
-
-        Args:
-            prompts (TensorDict): Input batch from train or validation dataset.
-        """
-        chunkes = prompts.chunk(len(self.agent_loop_workers))
-        ray.get(
-            [
-                worker.generate_sequences.remote(chunk)
-                for worker, chunk in zip(self.agent_loop_workers, chunkes, strict=False)
-            ]
-        )
-
-
-# ======================================= USER SECTION END =======================================
-
-
-class PPOTrainer:
-    """PPO Trainer with TransferQueue and ReplayBuffer.
-
-    Args:
-        config: DictConfig from yaml config file.
-        role_worker_mapping: dict[Role, WorkerType]
-        resource_pool_manager: ResourcePoolManager
-    """
-
-    def __init__(
-        self,
-        config: DictConfig,
-        role_worker_mapping: dict[Role, WorkerType],
-        resource_pool_manager: ResourcePoolManager,
-    ):
-        self.config = config
-        if config.algorithm.adv_estimator == core_algos.AdvantageEstimator.REMAX:
-            raise NotImplementedError("ReMax advantage estimator is not supported by the synchronous PPO trainer.")
-        self.role_worker_mapping = role_worker_mapping
-        self.resource_pool_manager = resource_pool_manager
-        self.use_critic = need_critic(self.config)
-        self.use_reference_policy = need_reference_policy(self.config)
-        self.use_teacher_policy = need_teacher_policy(self.config)
-        self.replay_buffer = ReplayBuffer(poll_interval=1.0)
-        if self.config.algorithm.use_kl_in_reward:
-            self.kl_ctrl_in_reward = core_algos.get_kl_controller(self.config.algorithm.kl_ctrl)
-
-        self._init_tokenizer()
-        self._init_dataloader()
-        self._init_dump_executor()
-
-    def _init_tokenizer(self):
-        """Initialize tokenizer."""
-        # Download the checkpoint from HDFS to the local machine.
-        # `use_shm` determines whether to use shared memory, which could lead to faster model loading if turned on
-        local_path = copy_to_local(
-            self.config.actor_rollout_ref.model.path, use_shm=self.config.actor_rollout_ref.model.get("use_shm", False)
-        )
-        trust_remote_code = self.config.data.get("trust_remote_code", False)
-        self.tokenizer = hf_tokenizer(local_path, trust_remote_code=trust_remote_code)
-        # Used for multimodal LLM, could be None
-        self.processor = hf_processor(local_path, trust_remote_code=trust_remote_code, use_fast=True)
-
-    def _init_dataloader(self):
-        """Initialize train and validate dataloader."""
-        self.train_dataset = create_rl_dataset(
-            self.config.data.train_files,
-            self.config.data,
-            self.tokenizer,
-            self.processor,
-            is_train=True,
-            max_samples=self.config.data.get("train_max_samples", -1),
-        )
-        self.val_dataset = create_rl_dataset(
-            self.config.data.val_files,
-            self.config.data,
-            self.tokenizer,
-            self.processor,
-            is_train=False,
-            max_samples=self.config.data.get("val_max_samples", -1),
-        )
-
-        self.train_dataloader = StatefulDataLoader(
-            dataset=self.train_dataset,
-            batch_size=self.config.data.get("gen_batch_size", self.config.data.train_batch_size),
-            num_workers=self.config.data["dataloader_num_workers"],
-            drop_last=True,
-            collate_fn=collate_fn,
-            sampler=create_rl_sampler(self.config.data, self.train_dataset),
-        )
-        self.val_dataloader = StatefulDataLoader(
-            dataset=self.val_dataset,
-            batch_size=self.config.data.val_batch_size or len(self.val_dataset),
-            num_workers=self.config.data["dataloader_num_workers"],
-            shuffle=self.config.data.get("validation_shuffle", True),
-            drop_last=False,
-            collate_fn=collate_fn,
-        )
-        logger.info(
-            f"train and validate dataloader initialized, train dataset size: "
-            f"{len(self.train_dataset)}, val dataset size: {len(self.val_dataset)}"
-        )
-
-        # adjust total_training_steps
-        total_training_steps = len(self.train_dataloader) * self.config.trainer.total_epochs
-        if self.config.trainer.total_training_steps is not None:
-            total_training_steps = self.config.trainer.total_training_steps
-        self.total_training_steps = total_training_steps
-        logger.info(f"Total training steps: {self.total_training_steps}")
-
-        try:
-            OmegaConf.set_struct(self.config, True)
-            with open_dict(self.config):
-                if OmegaConf.select(self.config, "actor_rollout_ref.actor.optim"):
-                    self.config.actor_rollout_ref.actor.optim.total_training_steps = total_training_steps
-                if OmegaConf.select(self.config, "critic.optim"):
-                    self.config.critic.optim.total_training_steps = total_training_steps
-        except Exception as e:
-            logger.warning(f"Warning: Could not set total_training_steps in config. Structure missing? Error: {e}")
-
-    def init_workers(self):
-        """Initialize distributed training workers using Ray backend.
-
-        Creates:
-        1. Ray resource pools from configuration
-        2. Worker groups for each role (actor, critic, etc.)
-        """
-        self.resource_pool_manager.create_resource_pool()
-        self.resource_pool_to_cls = {pool: {} for pool in self.resource_pool_manager.resource_pool_dict.values()}
-
-        # 1. define actor and rollout class
-        actor_role = Role.ActorRolloutRef if Role.ActorRolloutRef in self.role_worker_mapping else Role.ActorRollout
-        actor_rollout_resource_pool = self.resource_pool_manager.get_resource_pool(actor_role)
-        actor_rollout_cls = RayClassWithInitArgs(
-            cls=self.role_worker_mapping[actor_role],
-            config=self.config.actor_rollout_ref,
-            distillation_config=self.config.get("distillation"),
-            role=str(actor_role),
-        )
-        self.resource_pool_to_cls[actor_rollout_resource_pool][str(actor_role)] = actor_rollout_cls
-
-        # 2. define critic class
-        if self.use_critic:
-            critic_cfg: CriticConfig = omega_conf_to_dataclass(self.config.critic)
-            critic_cfg.engine.infer_max_token_len_per_gpu = critic_cfg.ppo_infer_max_token_len_per_gpu
-            critic_cfg.engine.max_token_len_per_gpu = critic_cfg.ppo_infer_max_token_len_per_gpu
-            worker_cfg = TrainingWorkerConfig(
-                model_type="value_model",
-                model_config=critic_cfg.model_config,
-                engine_config=critic_cfg.engine,
-                optimizer_config=critic_cfg.optim,
-                checkpoint_config=critic_cfg.checkpoint,
-            )
-            resource_pool = self.resource_pool_manager.get_resource_pool(Role.Critic)
-            critic_cls = RayClassWithInitArgs(cls=self.role_worker_mapping[Role.Critic], config=worker_cfg)
-            self.resource_pool_to_cls[resource_pool][str(Role.Critic)] = critic_cls
-
-        # 3. create worker group for actor rollout and critic
-        all_wg = {}
-        wg_kwargs = {}  # Setting up kwargs for RayWorkerGroup
-        if OmegaConf.select(self.config.global_profiler, "steps") is not None:
-            wg_kwargs["profile_steps"] = OmegaConf.select(self.config.global_profiler, "steps")
-            # Only require nsight worker options when tool is nsys
-            if OmegaConf.select(self.config.global_profiler, "tool") == "nsys":
-                assert (
-                    OmegaConf.select(self.config.global_profiler.global_tool_config.nsys, "worker_nsight_options")
-                    is not None
-                ), "worker_nsight_options must be set when using nsys with profile_steps"
-                wg_kwargs["worker_nsight_options"] = OmegaConf.to_container(
-                    OmegaConf.select(self.config.global_profiler.global_tool_config.nsys, "worker_nsight_options")
-                )
-        wg_kwargs["device_name"] = self.config.trainer.device
-        logger.info(f"worker group kwargs: {wg_kwargs}")
-
-        for resource_pool, class_dict in self.resource_pool_to_cls.items():
-            if not class_dict:
-                continue
-            worker_dict_cls = create_colocated_worker_cls(class_dict=class_dict)
-            wg_dict = RayWorkerGroup(
-                resource_pool=resource_pool,
-                ray_cls_with_init=worker_dict_cls,
-                **wg_kwargs,
-            )
-            spawn_wg = wg_dict.spawn(prefix_set=class_dict.keys())
-            all_wg.update(spawn_wg)
-            logger.info(f"create worker group {spawn_wg.keys()}")
-
-        # 5. initialize critic model engine
-        if self.use_critic:
-            self.critic_wg = all_wg[str(Role.Critic)]
-            self.critic_wg.reset()
-            value_loss_ = partial(value_loss, config=critic_cfg)
-            self.critic_wg.set_loss_fn(value_loss_)
-            logger.info("critic model engine initialized")
-
-        # 6. initialize actor and ref model engine
-        self.actor_rollout_wg = all_wg[str(actor_role)]
-        self.actor_rollout_wg.init_model()
-        logger.info("actor and ref model engine initialized")
-
-        # if ref_in_actor is True, the reference policy will be actor without lora applied
-        lora_rank = self.config.actor_rollout_ref.model.get("lora", {}).get("rank", 0)
-        if lora_rank <= 0:
-            lora_rank = self.config.actor_rollout_ref.model.get("lora_rank", 0)
-        self.ref_in_actor = lora_rank > 0 or self.config.actor_rollout_ref.model.get("lora_adapter_path") is not None
-        if self.use_reference_policy:
-            self.ref_policy_wg = all_wg[str(Role.ActorRolloutRef)]
-
-        # 7. initialize reward loop manager
-        resource_pool = (
-            self.resource_pool_manager.get_resource_pool(Role.RewardModel)
-            if self.config.reward.reward_model.enable
-            else None
-        )
-        self.reward_loop_manager = RewardLoopManager(
-            config=self.config,
-            rm_resource_pool=resource_pool,
-        )
-        logger.info("reward loop manager initialized")
-
-        # 8. initialize teacher loop manager
-        if self.use_teacher_policy:
-            teacher_resource_pool = self.resource_pool_manager.get_resource_pool(Role.TeacherModel)
-            self.teacher_model_manager = MultiTeacherModelManager(
-                config=self.config,
-                resource_pool=teacher_resource_pool,
-            )
-            self.distillation_config: DistillationConfig = omega_conf_to_dataclass(self.config.distillation)
-        else:
-            self.teacher_model_manager = None
-            self.distillation_config = None
-
-        # 9. initialize agent loop manager
-        self.llm_server_manager = LLMServerManager.create(
-            config=self.config, worker_group=self.actor_rollout_wg, rollout_resource_pool=actor_rollout_resource_pool
-        )
-
-        manager_class_fqn = self.config.actor_rollout_ref.rollout.get("agent", {}).get("agent_loop_manager_class")
-        if manager_class_fqn:
-            agent_loop_manager_cls = load_class_from_fqn(manager_class_fqn, "AgentLoopManager")
-        else:
-            agent_loop_manager_cls = AgentLoopManagerTQ
-        self.async_rollout_manager = agent_loop_manager_cls.create(
-            config=self.config,
-            llm_client=self.llm_server_manager.get_client(),
-            teacher_client=self.teacher_model_manager.get_client() if self.use_teacher_policy else None,
-            reward_loop_worker_handles=self.reward_loop_manager.reward_loop_workers,
-        )
-        logger.info("agent loop manager initialized")
-
-        # 10. initialize checkpoint engine manager
-        checkpoint_engine_config = omega_conf_to_dataclass(self.config.actor_rollout_ref.rollout.checkpoint_engine)
-        self.checkpoint_manager = CheckpointEngineManager(
-            config=checkpoint_engine_config,
-            trainer=self.actor_rollout_wg,
-            replicas=self.llm_server_manager.get_replicas(),
-        )
-        logger.info("checkpoint engine manager initialized")
-
-        # sleep all replicas to load checkpoint
-        self.checkpoint_manager.sleep_replicas()
-
-        logger.info("all initialize finished, ready to fit")
-
-    def _load_checkpoint(self):
-        self.global_steps = 0
-
-        # 1. find latest checkpoint folder
-        if self.config.trainer.resume_mode == "disable":
-            return
-        elif self.config.trainer.resume_mode == "auto":
-            checkpoint_folder = self.config.trainer.default_local_dir
-            if not os.path.isabs(checkpoint_folder):
-                working_dir = os.getcwd()
-                checkpoint_folder = os.path.join(working_dir, checkpoint_folder)
-            global_step_folder = find_latest_ckpt_path(checkpoint_folder)  # None if no latest
-            if global_step_folder is None:
-                logger.info("Training from scratch")
-                return
-        elif self.config.trainer.resume_mode == "resume_path":
-            assert isinstance(self.config.trainer.resume_from_path, str), "resume ckpt must be str type"
-            assert "global_step_" in self.config.trainer.resume_from_path, "resume ckpt must specify the global_steps"
-            global_step_folder = self.config.trainer.resume_from_path
-            if not os.path.isabs(global_step_folder):
-                working_dir = os.getcwd()
-                global_step_folder = os.path.join(working_dir, global_step_folder)
-        else:
-            logger.exception(f"Unknown resume mode {self.config.trainer.resume_mode}")
-
-        # set global step
-        self.global_steps = int(global_step_folder.split("global_step_")[-1])
-        logger.info(f"Resuming from {global_step_folder}, setting global step to {self.global_steps}")
-
-        # 2. load actor checkpoint
-        self.actor_rollout_wg.load_checkpoint(
-            local_path=os.path.join(global_step_folder, "actor"),
-            del_local_after_load=self.config.trainer.del_local_ckpt_after_load,
-        )
-
-        # 3. load critic checkpoint
-        if self.use_critic:
-            self.critic_wg.load_checkpoint(
-                local_path=os.path.join(global_step_folder, str(Role.Critic)),
-                del_local_after_load=self.config.trainer.del_local_ckpt_after_load,
-            )
-
-        # 4. load dataloader checkpoint
-        dataloader_local_path = os.path.join(global_step_folder, "data.pt")
-        if os.path.exists(dataloader_local_path):
-            dataloader_state_dict = torch.load(dataloader_local_path, weights_only=False)
-            self.train_dataloader.load_state_dict(dataloader_state_dict)
-        else:
-            logger.warning(f"Warning: No dataloader state found at {dataloader_local_path}, will start from scratch")
-
-    def _save_checkpoint(self):
-        """Save actor, critic, and dataloader checkpoints to local (and optionally remote) storage."""
-        from verl.utils.fs import local_mkdir_safe
-
-        local_global_step_folder = os.path.join(
-            self.config.trainer.default_local_dir, f"global_step_{self.global_steps}"
-        )
-        logger.info(f"Saving checkpoint to {local_global_step_folder}")
-
-        # resolve max checkpoints to keep
-        remove_previous_ckpt_in_save = self.config.trainer.get("remove_previous_ckpt_in_save", False)
-        if remove_previous_ckpt_in_save:
-            logger.warning(
-                "remove_previous_ckpt_in_save is deprecated, "
-                "set max_actor_ckpt_to_keep=1 and max_critic_ckpt_to_keep=1 instead"
-            )
-        max_actor_ckpt_to_keep = (
-            self.config.trainer.get("max_actor_ckpt_to_keep", None) if not remove_previous_ckpt_in_save else 1
-        )
-        max_critic_ckpt_to_keep = (
-            self.config.trainer.get("max_critic_ckpt_to_keep", None) if not remove_previous_ckpt_in_save else 1
-        )
-
-        # save actor
-        actor_local_path = os.path.join(local_global_step_folder, "actor")
-        actor_remote_path = (
-            None
-            if self.config.trainer.default_hdfs_dir is None
-            else os.path.join(self.config.trainer.default_hdfs_dir, f"global_step_{self.global_steps}", "actor")
-        )
-        self.actor_rollout_wg.save_checkpoint(
-            actor_local_path, actor_remote_path, self.global_steps, max_ckpt_to_keep=max_actor_ckpt_to_keep
-        )
-
-        # save critic
-        if self.use_critic:
-            critic_local_path = os.path.join(local_global_step_folder, str(Role.Critic))
-            critic_remote_path = (
-                None
-                if self.config.trainer.default_hdfs_dir is None
-                else os.path.join(
-                    self.config.trainer.default_hdfs_dir, f"global_step_{self.global_steps}", str(Role.Critic)
-                )
-            )
-            self.critic_wg.save_checkpoint(
-                critic_local_path, critic_remote_path, self.global_steps, max_ckpt_to_keep=max_critic_ckpt_to_keep
-            )
-
-        # save dataloader state
-        local_mkdir_safe(local_global_step_folder)
-        dataloader_local_path = os.path.join(local_global_step_folder, "data.pt")
-        torch.save(self.train_dataloader.state_dict(), dataloader_local_path)
-
-        # write latest checkpointed iteration tracker for atomic resume
-        actor_ckpt_cfg = self.config.actor_rollout_ref.actor.get("checkpoint", {})
-        if actor_ckpt_cfg.get("async_save", False):
-            logger.info("skip write latest_checkpointed_iteration.txt when async_save is True")
-            return
-        local_latest_checkpointed_iteration = os.path.join(
-            self.config.trainer.default_local_dir, "latest_checkpointed_iteration.txt"
-        )
-        with open(local_latest_checkpointed_iteration, "w") as f:
-            f.write(str(self.global_steps))
-
-    def _validate(self) -> dict[str, float]:
-        # Lists to collect samples for the table
-        sample_uids = []
-        sample_inputs = []
-        sample_outputs = []
-        sample_gts = []
-        sample_scores = []
-        sample_turns = []
-        data_sources = []
-        reward_extra_infos_dict: dict[str, list] = defaultdict(list)
-        dump_all_inputs: list[str] = []
-        dump_all_outputs: list[str] = []
-        dump_all_keys: list[str] = []
-        session_to_sample_idx: dict[str, int] = {}
-
-        for batch_dict in self.val_dataloader:
-            # 1. put batch to agent loop manager
-            batch_dict["uid"] = np.array(
-                [str(uuid.uuid4()) for _ in range(len(batch_dict["raw_prompt"]))], dtype=object
-            )
-            batch = tu.get_tensordict(batch_dict)
-            tu.assign_non_tensor_data(batch, "global_steps", self.global_steps)
-            tu.assign_non_tensor_data(batch, "validate", True)
-            # Register each prompt (GRPO group) in TransferQueue as a tag-only status marker
-            tags = [{"is_prompt": True, "status": "pending"}] * len(batch)
-            tq.kv_batch_put(keys=list(batch["uid"]), partition_id="val", tags=tags)
-            self.async_rollout_manager.generate_sequences(batch)
-
-            # 2. sample batch from replay buffer: one prompt (GRPO group) per submitted row.
-            batch = self.replay_buffer.sample(partition_id="val", batch_size=len(batch))
-
-            # 3. [OPTIONAL] compute reward score with colocated reward model
-            if self.reward_loop_manager.reward_loop_worker_handles is None:
-                self.checkpoint_manager.sleep_replicas()
-                batch = self._compute_reward_colocate(batch)
-                self.checkpoint_manager.update_weights()
-
-            # 4. collect necessary data for logging
-            # For multi-output agent loops, only use the final output per session for metrics.
-            # Keys have format {uid}_{session_id}_{index}; keep only the highest index per session.
-            session_max: dict[str, tuple[int, int]] = {}  # session_key -> (max_index, position)
-            for pos, key in enumerate(batch.keys):
-                parts = key.rsplit("_", 2)
-                if len(parts) == 3:
-                    session_key = f"{parts[0]}_{parts[1]}"
-                    index = int(parts[2])
-                    if session_key not in session_max or index > session_max[session_key][0]:
-                        session_max[session_key] = (index, pos)
-                else:
-                    session_max[key] = (0, pos)
-            sorted_sessions = sorted(session_max.items(), key=lambda x: x[1][1])
-            final_indices = [pos for _, (_, pos) in sorted_sessions]
-            final_keys = [batch.keys[i] for i in final_indices]
-            base_offset = len(sample_scores)
-            session_to_sample_idx.update(
-                {session_key: base_offset + j for j, (session_key, _) in enumerate(sorted_sessions)}
-            )
-
-            text_data = tq.kv_batch_get(
-                keys=batch.keys, partition_id=batch.partition_id, select_fields=["prompts", "responses"]
-            )
-            text_data["prompts"] = text_data["prompts"].to_padded_tensor(padding=self.tokenizer.pad_token_id)
-            text_data["responses"] = text_data["responses"].to_padded_tensor(padding=self.tokenizer.pad_token_id)
-            all_inputs = [self.tokenizer.decode(ids, skip_special_tokens=True) for ids in text_data["prompts"]]
-            all_outputs = [self.tokenizer.decode(ids, skip_special_tokens=True) for ids in text_data["responses"]]
-
-            fields = ["uid", "rm_scores", "num_turns", "reward_model", "data_source", "extra_fields"]
-            data = tq.kv_batch_get(keys=final_keys, partition_id=batch.partition_id, select_fields=fields)
-
-            sample_uids.extend(data.pop("uid").tolist())
-            sample_outputs.extend(all_outputs[i] for i in final_indices)
-            sample_inputs.extend(all_inputs[i] for i in final_indices)
-            scores = data["rm_scores"].sum(dim=1).tolist()
-            sample_scores.extend(scores)
-            sample_turns.extend(data.pop("num_turns").tolist())
-            reward_extra_infos_dict["reward"].extend(scores)
-
-            extra_fields_list = data.pop("extra_fields", None)
-            if extra_fields_list is not None:
-                n_prior = len(reward_extra_infos_dict["reward"]) - len(extra_fields_list.tolist())
-                for extra_field in extra_fields_list.tolist():
-                    reward_extra_info = (
-                        extra_field.get("reward_extra_info", {}) if isinstance(extra_field, dict) else {}
-                    )
-                    for key in reward_extra_infos_dict:
-                        if key != "reward" and key not in reward_extra_info:
-                            reward_extra_infos_dict[key].append(None)
-                    for key, value in reward_extra_info.items():
-                        if key not in reward_extra_infos_dict:
-                            reward_extra_infos_dict[key] = [None] * n_prior
-                        reward_extra_infos_dict[key].append(value)
-                    n_prior += 1
-
-            reward_model = data.pop("reward_model", None)
-            if reward_model is not None:
-                sample_gts.extend([item.get("ground_truth", None) for item in reward_model.tolist()])
-            else:
-                sample_gts.extend([None] * len(final_indices))
-
-            data_source = data.pop("data_source", None)
-            if data_source is not None:
-                data_sources.extend(data_source.tolist())
-            else:
-                data_sources.extend(["unknown"] * len(final_indices))
-
-            dump_all_inputs.extend(all_inputs)
-            dump_all_outputs.extend(all_outputs)
-            dump_all_keys.extend(batch.keys)
-
-            # 5. cleanup transfer queue
-            tq.kv_clear(keys=batch.keys, partition_id=batch.partition_id)
-
-        # logger to wandb
-        self._maybe_log_val_generations(inputs=sample_inputs, outputs=sample_outputs, scores=sample_scores)
-
-        # dump to local dir
-        val_data_dir = self.config.trainer.get("validation_data_dir", None)
-        if val_data_dir:
-            # Sort according to uid (so that generations in the same rollout are together)
-            sort_keys = []
-            for key in dump_all_keys:
-                parts = key.rsplit("_", 2)
-                sort_keys.append((parts[0], int(parts[1]), int(parts[2])) if len(parts) == 3 else (key, 0, 0))
-            sorted_indices = sorted(range(len(dump_all_keys)), key=lambda i: sort_keys[i])
-            dump_all_inputs = [dump_all_inputs[i] for i in sorted_indices]
-            dump_all_outputs = [dump_all_outputs[i] for i in sorted_indices]
-            dump_all_keys = [dump_all_keys[i] for i in sorted_indices]
-
-            # For ground truths, scores and reward extra infos, find the values in the
-            # lists for the final samples of each session
-            dump_all_sessions = [
-                f"{parts[0]}_{parts[1]}" if len(parts) == 3 else key
-                for key in dump_all_keys
-                for parts in [key.rsplit("_", 2)]
-            ]
-            session_final_indices = [session_to_sample_idx[session] for session in dump_all_sessions]
-            self._dump_generations(
-                inputs=dump_all_inputs,
-                outputs=dump_all_outputs,
-                gts=[sample_gts[i] for i in session_final_indices],
-                scores=[sample_scores[i] for i in session_final_indices],
-                reward_extra_infos_dict={
-                    k: [v[i] for i in session_final_indices] for k, v in reward_extra_infos_dict.items()
-                }
-                | {"uid": dump_all_keys},
-                dump_path=val_data_dir,
-            )
-
-        return self._val_metrics_update(data_sources, sample_uids, reward_extra_infos_dict, sample_turns)
-
-    def _maybe_log_val_generations(self, inputs, outputs, scores):
-        """Log a table of validation samples to the configured logger (wandb or swanlab)"""
-        generations_to_log = self.config.trainer.log_val_generations
-        if generations_to_log == 0:
-            return
-
-        # Create tuples of (input, output, score) and sort by input text
-        samples = list(zip(inputs, outputs, scores, strict=True))
-        samples.sort(key=lambda x: x[0])  # Sort by input text
-
-        # Use fixed random seed for deterministic shuffling
-        rng = np.random.RandomState(42)
-        rng.shuffle(samples)
-
-        # Take first N samples after shuffling
-        samples = samples[:generations_to_log]
-
-        # Log to each configured logger
-        self.validation_generations_logger.log(self.config.trainer.logger, samples, self.global_steps)
-
-    @staticmethod
-    def _write_generations(inputs, outputs, gts, scores, reward_extra_infos_dict, dump_path, global_steps):
-        """Write generation samples as JSONL (runs in background thread)."""
-        os.makedirs(dump_path, exist_ok=True)
-        filename = os.path.join(dump_path, f"{global_steps}.jsonl")
-
-        n = len(inputs)
-        base_data = {
-            "input": inputs,
-            "output": outputs,
-            "gts": gts,
-            "score": scores,
-            "step": [global_steps] * n,
-        }
-
-        for k, v in reward_extra_infos_dict.items():
-            if len(v) == n:
-                base_data[k] = v
-
-        def json_encode_default(obj):
-            if isinstance(obj, np.integer):
-                return int(obj)
-            elif isinstance(obj, np.floating):
-                return float(obj)
-            elif isinstance(obj, np.bool_):
-                return bool(obj)
-            elif hasattr(obj, "tolist"):
-                return obj.tolist()
-            raise TypeError(f"Object of type {type(obj).__name__} is not JSON serializable")
-
-        with open(filename, "w") as f:
-            for i in range(n):
-                entry = {k: v[i] for k, v in base_data.items()}
-                f.write(json.dumps(entry, ensure_ascii=False, default=json_encode_default) + "\n")
-
-        print(f"Dumped generations to {filename}")
-
-    def _dump_generations(self, inputs, outputs, gts, scores, reward_extra_infos_dict, dump_path):
-        """Dump rollout/validation samples as JSONL asynchronously."""
-        global_steps = self.global_steps
-        future = self._dump_executor.submit(
-            self._write_generations,
-            inputs,
-            outputs,
-            gts,
-            scores,
-            reward_extra_infos_dict,
-            dump_path,
-            global_steps,
-        )
-        self._dump_futures.append(future)
-        # Clean up completed futures and surface any exceptions early
-        still_pending = []
-        for f in self._dump_futures:
-            if f.done():
-                f.result()  # re-raises if the write failed
-            else:
-                still_pending.append(f)
-        self._dump_futures = still_pending
-
-    def _init_dump_executor(self):
-        """Create or recreate the dump executor and futures list."""
-        self._dump_executor = ThreadPoolExecutor(max_workers=1)
-        self._dump_futures = []
-
-    def _shutdown_dump_executor(self):
-        """Drain pending dump futures and shut down the executor."""
-        for f in self._dump_futures:
-            f.result()
-        self._dump_futures.clear()
-        self._dump_executor.shutdown(wait=True)
-
-    def _log_rollout_data(self, batch: KVBatchMeta, timing_raw: dict, rollout_data_dir: str):
-        """Fetch rollout data from TransferQueue and dump sorted by uid."""
-        with marked_timer("dump_rollout_generations", timing_raw, color="green"):
-            fields = ["uid", "prompts", "responses", "rm_scores", "reward_model"]
-            data = tq.kv_batch_get(keys=batch.keys, partition_id=batch.partition_id, select_fields=fields)
-            data["prompts"] = data["prompts"].to_padded_tensor(padding=self.tokenizer.pad_token_id)
-            data["responses"] = data["responses"].to_padded_tensor(padding=self.tokenizer.pad_token_id)
-
-            uids = data.pop("uid").tolist()
-            inputs = [self.tokenizer.decode(ids, skip_special_tokens=True) for ids in data["prompts"]]
-            outputs = [self.tokenizer.decode(ids, skip_special_tokens=True) for ids in data["responses"]]
-            scores = data["rm_scores"].sum(dim=1).tolist()
-
-            reward_model = data.pop("reward_model", None)
-            if reward_model is not None:
-                gts = [item.get("ground_truth", None) for item in reward_model.tolist()]
-            else:
-                gts = [None] * len(uids)
-
-            # Sort by uid key ({sample}_{rollout}_{output})
-            sort_keys = []
-            for key in batch.keys:
-                parts = key.rsplit("_", 2)
-                if len(parts) == 3:
-                    sort_keys.append((parts[0], int(parts[1]), int(parts[2])))
-                else:
-                    sort_keys.append((key, 0, 0))
-            sorted_indices = sorted(range(len(sort_keys)), key=lambda i: sort_keys[i])
-
-            inputs = [inputs[i] for i in sorted_indices]
-            outputs = [outputs[i] for i in sorted_indices]
-            gts = [gts[i] for i in sorted_indices]
-            scores = [scores[i] for i in sorted_indices]
-
-            reward_extra_infos_dict = {"uid": [batch.keys[i] for i in sorted_indices]}
-
-            self._dump_generations(
-                inputs=inputs,
-                outputs=outputs,
-                gts=gts,
-                scores=scores,
-                reward_extra_infos_dict=reward_extra_infos_dict,
-                dump_path=rollout_data_dir,
-            )
-
-    def _val_metrics_update(self, data_sources, sample_uids, reward_extra_infos_dict, sample_turns) -> dict[str, float]:
-        data_src2var2metric2val = process_validation_metrics(data_sources, sample_uids, reward_extra_infos_dict)
-        metric_dict = {}
-        for data_source, var2metric2val in data_src2var2metric2val.items():
-            core_var = "acc" if "acc" in var2metric2val else "reward"
-            for var_name, metric2val in var2metric2val.items():
-                n_max = max([int(name.split("@")[-1].split("/")[0]) for name in metric2val.keys()])
-                for metric_name, metric_val in metric2val.items():
-                    if (
-                        (var_name == core_var)
-                        and any(metric_name.startswith(pfx) for pfx in ["mean", "maj", "best"])
-                        and (f"@{n_max}" in metric_name)
-                    ):
-                        metric_sec = "val-core"
-                    else:
-                        metric_sec = "val-aux"
-                    pfx = f"{metric_sec}/{data_source}/{var_name}/{metric_name}"
-                    metric_dict[pfx] = metric_val
-
-        if len(sample_turns) > 0:
-            sample_turns = np.array(sample_turns)
-            metric_dict["val-aux/num_turns/min"] = sample_turns.min()
-            metric_dict["val-aux/num_turns/max"] = sample_turns.max()
-            metric_dict["val-aux/num_turns/mean"] = sample_turns.mean()
-
-        return metric_dict
-
-    def _start_profiling(self) -> None:
-        """Start profiling for all worker groups if profiling is enabled."""
-        do_profile = (
-            not self.prev_step_profile and self.curr_step_profile
-            if self.config.global_profiler.profile_continuous_steps
-            else self.curr_step_profile
-        )
-
-        if do_profile:
-            self.actor_rollout_wg.start_profile(role="e2e", profile_step=self.global_steps)
-            if self.use_reference_policy:
-                self.ref_policy_wg.start_profile(profile_step=self.global_steps)
-            if self.use_critic:
-                self.critic_wg.start_profile(profile_step=self.global_steps)
-
-    def _stop_profiling(self) -> None:
-        """Stop profiling for all worker groups if profiling is enabled."""
-        self.next_step_profile = (
-            self.global_steps + 1 in self.config.global_profiler.steps
-            if self.config.global_profiler.steps is not None
-            else False
-        )
-        do_profile = (
-            self.curr_step_profile and not self.next_step_profile
-            if self.config.global_profiler.profile_continuous_steps
-            else self.curr_step_profile
-        )
-        self.prev_step_profile = self.curr_step_profile
-        self.curr_step_profile = self.next_step_profile
-
-        if do_profile:
-            self.actor_rollout_wg.stop_profile()
-            if self.use_reference_policy:
-                self.ref_policy_wg.stop_profile()
-            if self.use_critic:
-                self.critic_wg.stop_profile()
-
-    def _compute_reward_colocate(self, batch: KVBatchMeta, metrics: dict) -> KVBatchMeta:
-        """Compute the reward with colocate reward model."""
-        # TODO: add reward model
-        raise NotImplementedError
-
-    def _get_required_batch_multiple(self, dp_size: int) -> int:
-        """Return the global batch multiple required by downstream train steps(e.g. critics, actors)."""
-        required_multiple = dp_size
-
-        # If enabled with critic training, the batch should align with critic PPO mini-batches.
-        if self.use_critic:
-            critic_global_mini_batch_size = self.config.critic.ppo_mini_batch_size
-            critic_global_mini_batch_size *= self.config.actor_rollout_ref.rollout.n
-            required_multiple = math.lcm(required_multiple, critic_global_mini_batch_size)
-
-        # If there is an actor update, the batch should align with actor PPO mini-batches too.
-        if self.config.trainer.critic_warmup <= self.global_steps:
-            actor_global_mini_batch_size = self.config.actor_rollout_ref.actor.ppo_mini_batch_size
-            actor_global_mini_batch_size *= self.config.actor_rollout_ref.rollout.n
-            required_multiple = math.lcm(required_multiple, actor_global_mini_batch_size)
-
-        # Notice lcm(a, b, c) == lcm(lcm(a, b), c), so it is optimal.
-        return required_multiple
-
-    def _balance_batch(self, batch: KVBatchMeta, metrics, logging_prefix="global_seqlen", keep_minibatch=False):
-        """Reorder the data on single controller such that each dp rank gets similar total tokens."""
-        # get actor dp size
-        role, worker_group = "actor", self.actor_rollout_wg
-        if role not in worker_group._dispatch_info:
-            dp_rank_mapping = worker_group._query_dispatch_info(role)
-            worker_group._dispatch_info[role] = dp_rank_mapping
-        else:
-            dp_rank_mapping = worker_group._dispatch_info[role]
-        dp_size = max(dp_rank_mapping) + 1
-
-        # Upsampling the batch with padding sequences
-        batch_multiple = self._get_required_batch_multiple(dp_size)
-        batch = upsample_batch_to_divisible_size(batch, batch_multiple, self.tokenizer.eos_token_id)
-        global_seqlen_lst = torch.tensor([tag["seq_len"] for tag in batch.tags], dtype=torch.int64)
-        workload_lst = calculate_workload(global_seqlen_lst)
-
-        # reorder based on index. The data will be automatically equally partitioned by dispatch function
-        global_partition_lst = get_seqlen_balanced_partitions(workload_lst, k_partitions=dp_size, equal_size=True)
-        batch.reorder([j for partition in global_partition_lst for j in partition])
-        global_balance_stats = log_seqlen_unbalance(
-            seqlen_list=global_seqlen_lst.tolist(), partitions=global_partition_lst, prefix=logging_prefix
-        )
-        metrics.update(global_balance_stats)
-        return batch
-
-    def _compute_old_log_prob(self, batch: KVBatchMeta, metrics: dict) -> KVBatchMeta:
-        """Compute the old log prob of the batch."""
-        # Operating Mode Selection:
-        # - Bypass mode: Sets old_log_probs = rollout_log_probs (2 policies: π_rollout, π_θ)
-        # - Decoupled mode: Recomputes old_log_probs as proximal anchor (3 policies: π_rollout, π_old, π_θ)
-        #   Note: π_old computed once per data batch, serves as stable reference during mini-batch updates
-        rollout_corr_config = self.config.algorithm.get("rollout_correction", None)
-        bypass_recomputing_logprobs = rollout_corr_config and rollout_corr_config.get("bypass_mode", False)
-        if bypass_recomputing_logprobs:  # Use `rollout_log_probs`
-            data = tq.kv_batch_get(
-                keys=batch.keys, partition_id=batch.partition_id, select_fields=["rollout_log_probs"]
-            )
-            data["old_log_probs"] = data.pop("rollout_log_probs")
-            tq.kv_batch_put(keys=batch.keys, partition_id=batch.partition_id, fields=data)
-            return
-
-        # 1. compute log probs
-        batch.extra_info.update(
-            {
-                "calculate_entropy": True,
-                "compute_loss": False,
-                "temperature": self.config.actor_rollout_ref.rollout.temperature,
-            }
-        )
-        output: KVBatchMeta = self.actor_rollout_wg.compute_log_prob(batch)
-        assert len(output) == len(batch)
-
-        fields = ["entropy", "log_probs", "response_mask"]
-        if self.config.actor_rollout_ref.rollout.calculate_log_probs:
-            fields.extend(["responses", "rollout_log_probs"])
-        data = tq.kv_batch_get(keys=batch.keys, partition_id=batch.partition_id, select_fields=fields)
-
-        # 2. write old_log_probs and entropy back to TransferQueue
-        data["old_log_probs"] = response_from_nested(data.pop("log_probs"), data["response_mask"])
-        data["entropy"] = response_from_nested(data.pop("entropy"), data["response_mask"])
-        batch = tq.kv_batch_put(
-            keys=batch.keys, partition_id=batch.partition_id, fields=data.select("old_log_probs", "entropy")
-        )
-
-        data = DataProto(batch=data.to_padded_tensor())
-
-        # 3. calculate actor entroy metrics
-        actor_config = self.config.actor_rollout_ref.actor
-        entropy_agg = agg_loss(
-            loss_mat=data.batch["entropy"],
-            loss_mask=data.batch["response_mask"],
-            loss_agg_mode=actor_config.loss_agg_mode,
-            loss_scale_factor=actor_config.loss_scale_factor,
-        )
-        old_log_prob_metrics = {
-            "actor/entropy": entropy_agg.detach().item(),
-            # "perf/mfu/actor_infer": old_log_prob_mfu,
-        }
-        metrics.update(old_log_prob_metrics)
-
-        # 4. calculate rollout vs actor logprobs diff
-        if self.config.actor_rollout_ref.rollout.calculate_log_probs:
-            metrics.update(calculate_debug_metrics(data))
-
-        return batch
-
-    def _compute_ref_log_prob(self, batch: KVBatchMeta, metrics: dict) -> KVBatchMeta:
-        """Compute the reference log prob of the batch."""
-        # 1. compute log probs
-        metadata = {
-            "calculate_entropy": False,
-            "compute_loss": False,
-            "temperature": self.config.actor_rollout_ref.rollout.temperature,
-        }
-        if self.ref_in_actor:
-            metadata["no_lora_adapter"] = True
-        batch.extra_info.update(metadata)
-        if self.ref_in_actor:
-            output = self.actor_rollout_wg.compute_log_prob(batch)
-        else:
-            output = self.ref_policy_wg.compute_ref_log_prob(batch)
-        assert len(output) == len(batch)
-
-        # 2. write ref_log_prob and entropy back to TransferQueue
-        data = tq.kv_batch_get(
-            keys=batch.keys, partition_id=batch.partition_id, select_fields=["log_probs", "response_mask"]
-        )
-        data["ref_log_prob"] = response_from_nested(data.pop("log_probs"), data["response_mask"])
-        tq.kv_batch_put(keys=batch.keys, partition_id=batch.partition_id, fields=data.select("ref_log_prob"))
-
-        return batch
-
-    def _compute_values(self, batch: KVBatchMeta, metrics: dict) -> KVBatchMeta:
-        """Compute the values of the batch."""
-        # 1. compute value
-        output = self.critic_wg.infer_batch(batch)
-        # TODO: DataProtoFuture support KVBatchMeta
-        ray.get(output.futures)
-
-        # 2. write value back to TransferQueue
-        data = tq.kv_batch_get(
-            keys=batch.keys, partition_id=batch.partition_id, select_fields=["values", "response_mask"]
-        )
-        data["values"] = response_from_nested(data.pop("values"), data["response_mask"])
-        tq.kv_batch_put(keys=batch.keys, partition_id=batch.partition_id, fields=data.select("values"))
-
-        return batch
-
-    def _compute_advantage(self, batch: KVBatchMeta, metrics: dict) -> KVBatchMeta:
-        """Compute the advantage of the batch."""
-        fields = ["uid", "response_mask", "rm_scores", "rollout_log_probs", "old_log_probs", "ref_log_prob", "values"]
-        data = tq.kv_batch_get(keys=batch.keys, partition_id=batch.partition_id, select_fields=fields)
-
-        response_mask = data["response_mask"]
-        data = DataProto(batch=data.to_padded_tensor())
-        data.batch["token_level_scores"] = data.batch["rm_scores"]
-        data.non_tensor_batch["uid"] = np.array(data.batch.pop("uid").tolist(), dtype=object)
-
-        # 1. apply kl penalty to rewards
-        if self.config.algorithm.use_kl_in_reward:
-            data, kl_metrics = apply_kl_penalty(
-                data, kl_ctrl=self.kl_ctrl_in_reward, kl_penalty=self.config.algorithm.kl_penalty
-            )
-            metrics.update(kl_metrics)
-        else:
-            data.batch["token_level_rewards"] = data.batch["token_level_scores"]
-
-        # 2. Compute rollout correction: IS weights, rejection sampling, and metrics
-        # Only runs in decoupled mode (computes once per batch using stable π_old)
-        # In bypass mode, this is skipped - actor computes metrics from evolving π_θ vs π_rollout
-        rollout_corr_config = self.config.algorithm.get("rollout_correction", None)
-        bypass_recomputing_logprobs = rollout_corr_config and rollout_corr_config.get("bypass_mode", False)
-        rollout_correction = (
-            rollout_corr_config is not None and "rollout_log_probs" in data.batch and not bypass_recomputing_logprobs
-        )
-        if rollout_correction:
-            data, is_metrics = compute_rollout_correction_and_add_to_batch(data, rollout_corr_config)
-            metrics.update(is_metrics)
-
-        # 3. compute advantages
-        data = compute_advantage_for_multi_trajectories(
-            data,
-            batch_keys=batch.keys,
-            adv_estimator=self.config.algorithm.adv_estimator,
-            gamma=self.config.algorithm.gamma,
-            lam=self.config.algorithm.lam,
-            num_repeat=self.config.actor_rollout_ref.rollout.n,
-            norm_adv_by_std_in_grpo=self.config.algorithm.get("norm_adv_by_std_in_grpo", True),
-            config=self.config.algorithm,
-        )
-
-        # 4. write nested advantages and returns back to TransferQueue
-        fields = ["advantages", "returns"]
-        if self.config.algorithm.use_kl_in_reward:
-            fields.append("token_level_rewards")
-        if rollout_correction:
-            fields.append("response_mask")
-            if "rollout_is_weights" in data.batch:
-                fields.append("rollout_is_weights")
-
-        output = {}
-        for field in fields:
-            output[field] = response_to_nested(data.batch[field], response_mask)
-        output = TensorDict(output, batch_size=len(batch))
-
-        batch = tq.kv_batch_put(keys=batch.keys, partition_id=batch.partition_id, fields=output)
-
-        return batch
-
-    def _update_critic(self, batch: KVBatchMeta, metrics: dict) -> KVBatchMeta:
-        """Update the critic network."""
-        ppo_mini_batch_size = self.config.critic.ppo_mini_batch_size
-        ppo_mini_batch_size = ppo_mini_batch_size * self.config.actor_rollout_ref.rollout.n
-        extra_info = {
-            "global_batch_size": ppo_mini_batch_size,
-            "mini_batch_size": ppo_mini_batch_size,
-            "epochs": self.config.critic.ppo_epochs,
-            "seed": self.config.critic.data_loader_seed,
-            "dataloader_kwargs": {"shuffle": self.config.critic.shuffle},
-        }
-        batch.extra_info.update(extra_info)
-
-        output: DataProtoFuture = self.critic_wg.train_mini_batch(batch)
-        output: TensorDict = output.get()
-        output = rename_dict(output["metrics"], "critic/")
-        output["perf/mfu/critic"] = output.pop("critic/mfu")
-        critic_metrics = reduce_metrics(output)
-        metrics.update(critic_metrics)
-
-        return batch
-
-    def _update_actor(self, batch: KVBatchMeta, metrics: dict) -> KVBatchMeta:
-        """Update the actor network."""
-        ppo_mini_batch_size = self.config.actor_rollout_ref.actor.ppo_mini_batch_size
-        ppo_mini_batch_size = ppo_mini_batch_size * self.config.actor_rollout_ref.rollout.n
-        calculate_entropy = self.config.actor_rollout_ref.actor.calculate_entropy or (
-            self.config.actor_rollout_ref.actor.entropy_coeff != 0.0
-        )
-        distillation_use_topk = (
-            self.distillation_config.distillation_loss.loss_settings.use_topk
-            if is_distillation_enabled(self.config.get("distillation"))
-            else False
-        )
-        extra_info = {
-            "calculate_entropy": calculate_entropy,
-            "distillation_use_topk": distillation_use_topk,
-            "global_batch_size": ppo_mini_batch_size,
-            "mini_batch_size": ppo_mini_batch_size,
-            "epochs": self.config.actor_rollout_ref.actor.ppo_epochs,
-            "seed": self.config.actor_rollout_ref.actor.data_loader_seed,
-            "dataloader_kwargs": {"shuffle": self.config.actor_rollout_ref.actor.shuffle},
-            "temperature": self.config.actor_rollout_ref.rollout.temperature,
-        }
-        batch.extra_info.update(extra_info)
-
-        output: TensorDict = self.actor_rollout_wg.update_actor(batch)
-        output = rename_dict(output["metrics"], "actor/")
-        output["perf/mfu/actor"] = output.pop("actor/mfu")
-        actor_metrics = reduce_metrics(output)
-        metrics.update(actor_metrics)
-
-        return batch
-
-    def _compute_metrics(self, batch: KVBatchMeta, metrics, timing_raw, global_steps, epoch):
-        # 1. collect necessary fields from TransferQueue for computing metrics
-        non_padding_mask = np.array([not tag.get("is_padding", False) for tag in batch.tags], dtype=bool)
-        fields = [
-            "prompts",
-            "responses",
-            "response_mask",
-            "values",
-            "advantages",
-            "returns",
-            "rm_scores",
-            "token_level_rewards",
-            "num_turns",
-        ]
-        data = tq.kv_batch_get(keys=batch.keys, partition_id=batch.partition_id, select_fields=fields)
-        num_turns = np.array(data.pop("num_turns").tolist())
-        prompt_length = data["prompts"].offsets().diff()
-        response_length = data["responses"].offsets().diff()
-        global_token_num = (prompt_length + response_length).tolist()
-
-        # Only fetch speculative decoding stats when rollout writes them.
-        spec_drafts = spec_accepts = spec_verifies = None
-        mtp_config = getattr(self.config.actor_rollout_ref.model, "mtp", None)
-        if mtp_config is not None and mtp_config.enable and mtp_config.enable_rollout:
-            spec_data = tq.kv_batch_get(
-                keys=batch.keys,
-                partition_id=batch.partition_id,
-                select_fields=["extra_fields"],
-            )
-            extra_fields = spec_data["extra_fields"].tolist()
-            spec_drafts = [extra_field["spec_num_draft_tokens"] for extra_field in extra_fields]
-            spec_accepts = [extra_field["spec_num_accepted_tokens"] for extra_field in extra_fields]
-            spec_verifies = [extra_field["spec_num_verify_steps"] for extra_field in extra_fields]
-
-        data = data.to_padded_tensor()
-        data["token_level_scores"] = data["rm_scores"]
-        if "token_level_rewards" not in data:
-            data["token_level_rewards"] = data["rm_scores"]
-        data["prompt_length"] = prompt_length.float()
-        data["response_length"] = response_length.float()
-        batch = DataProto(batch=data, meta_info={"global_token_num": global_token_num})
-        metrics_batch = batch.select_idxs(non_padding_mask) if non_padding_mask.any() else batch
-
-        # 2. compute metrics
-        metrics.update({"training/global_step": global_steps, "training/epoch": epoch})
-        metrics.update(compute_data_metrics(batch=metrics_batch, use_critic=self.use_critic))
-        metrics.update(compute_timing_metrics(batch=batch, timing_raw=timing_raw))
-        n_gpus = self.resource_pool_manager.get_n_gpus()
-        metrics.update(compute_throughout_metrics(batch=batch, timing_raw=timing_raw, n_gpus=n_gpus))
-        gradient_norm = metrics.get("actor/grad_norm", None)
-        metrics.update(compute_variance_proxy_metrics(batch=metrics_batch, gradient_norm=gradient_norm))
-
-        # 3. other auxiliary metrics
-        if non_padding_mask.any():
-            num_turns = num_turns[non_padding_mask]
-        metrics.update(
-            {
-                "training/num_turns/mean": num_turns.mean(),
-                "training/num_turns/max": num_turns.max(),
-                "training/num_turns/min": num_turns.min(),
-            }
-        )
-
-        # 4. per-request speculative-decoding aggregation (same metrics async PPO logs;
-        # see compute_spec_decode_metrics in verl/trainer/ppo/ray_trainer.py).
-        metrics.update(compute_spec_decode_metrics(spec_drafts, spec_accepts, spec_verifies, non_padding_mask))
-
-    def fit(self):
-        if self._dump_executor._shutdown:
-            self._init_dump_executor()
-
-        self.logger = Tracking(
-            project_name=self.config.trainer.project_name,
-            experiment_name=self.config.trainer.experiment_name,
-            default_backend=self.config.trainer.logger,
-            config=OmegaConf.to_container(self.config, resolve=True),
-        )
-        self.validation_generations_logger = ValidationGenerationsLogger(
-            project_name=self.config.trainer.project_name,
-            experiment_name=self.config.trainer.experiment_name,
-        )
-
-        # load checkpoint and update weights before doing anything
-        self._load_checkpoint()
-        self.checkpoint_manager.update_weights()
-
-        # perform validation before training
-        if self.config.trainer.get("val_before_train", True):
-            val_metrics = self._validate()
-            assert val_metrics, f"{val_metrics=}"
-            pprint(f"Initial validation metrics: {val_metrics}")
-            self.logger.log(data=val_metrics, step=self.global_steps)
-            if self.config.trainer.get("val_only", False):
-                self._shutdown_dump_executor()
-                return
-
-        current_epoch = self.global_steps // len(self.train_dataloader)
-        progress_bar = tqdm(total=self.total_training_steps, initial=self.global_steps, desc="Training Progress")
-
-        # we start from step 1
-        self.global_steps += 1
-        self.prev_step_profile = False
-        self.curr_step_profile = (
-            self.global_steps in self.config.global_profiler.steps
-            if self.config.global_profiler.steps is not None
-            else False
-        )
-        self.next_step_profile = False
-
-        last_val_metrics = None
-        for epoch in range(current_epoch, self.config.trainer.total_epochs):
-            for batch_dict in self.train_dataloader:
-                is_last_step = self.global_steps >= self.total_training_steps
-                metrics, timing_raw = {}, {}
-
-                # 1. perform rollout and actor/critic training
-                self._start_profiling()
-                with marked_timer("step", timing_raw):
-                    batch = self.step(batch_dict, metrics, timing_raw)
-
-                    # 2. save checkpoint
-                    if self.config.trainer.save_freq > 0 and (
-                        is_last_step or self.global_steps % self.config.trainer.save_freq == 0
-                    ):
-                        with marked_timer("save_checkpoint", timing_raw, color="green"):
-                            self._save_checkpoint()
-
-                    # 3. update weights from trainer to rollout
-                    with marked_timer("update_weights", timing_raw, color="red"):
-                        self.checkpoint_manager.update_weights()
-                self._stop_profiling()
-
-                # 4. validate
-                if self.config.trainer.test_freq > 0 and (
-                    is_last_step or self.global_steps % self.config.trainer.test_freq == 0
-                ):
-                    with marked_timer("testing", timing_raw, color="green"):
-                        val_metrics: dict = self._validate()
-                        if is_last_step:
-                            last_val_metrics = val_metrics
-                    metrics.update(val_metrics)
-
-                # 5. record metrics
-                self._compute_metrics(batch, metrics, timing_raw, global_steps=self.global_steps, epoch=epoch)
-
-                # 6. dump rollout generations if enabled
-                rollout_data_dir = self.config.trainer.get("rollout_data_dir", None)
-                if rollout_data_dir:
-                    self._log_rollout_data(batch, timing_raw, rollout_data_dir)
-
-                # 7. cleanup transfer queue
-                tq.kv_clear(keys=batch.keys, partition_id=batch.partition_id)
-
-                self.logger.log(data=metrics, step=self.global_steps)
-                progress_bar.update(1)
-                self.global_steps += 1
-                if is_last_step:
-                    self._shutdown_dump_executor()
-                    pprint(f"Final validation metrics: {last_val_metrics}")
-                    progress_bar.close()
-                    return
-
-        # Ensure dump executor is shut down when training loop ends without reaching is_last_step
-        self._shutdown_dump_executor()
-
-    def step(self, batch_dict: dict, metrics: dict, timing_raw: dict) -> KVBatchMeta:
-        # 1. put batch to agent loop manager
-        batch_dict["uid"] = np.array([str(uuid.uuid4()) for _ in range(len(batch_dict["raw_prompt"]))], dtype=object)
-        batch = tu.get_tensordict(batch_dict)
-        tu.assign_non_tensor_data(batch, "global_steps", self.global_steps)
-        # Register each prompt (GRPO group) in TransferQueue as a tag-only status marker
-        tags = [{"is_prompt": True, "status": "pending"}] * len(batch)
-        tq.kv_batch_put(keys=list(batch["uid"]), partition_id="train", tags=tags)
-        self.async_rollout_manager.generate_sequences(batch)
-
-        # 2. sample batch from replay buffer
-        with marked_timer("gen", timing_raw, color="red"):
-            batch = self.replay_buffer.sample(partition_id="train", batch_size=len(batch))
-        batch.extra_info["temperature"] = self.config.actor_rollout_ref.rollout.temperature
-        self.checkpoint_manager.sleep_replicas()
-
-        # 3. [OPTIONAL] compute reward score with colocated reward model
-        if self.reward_loop_manager.reward_loop_worker_handles is None:
-            with marked_timer("reward", timing_raw, color="yellow"):
-                batch = self._compute_reward_colocate(batch)
-
-        # 4. balance batch across data parallel groups
-        batch = self._balance_batch(batch, metrics=metrics)
-
-        # 5. compute old_log_prob
-        with marked_timer("old_log_prob", timing_raw, color="blue"):
-            batch = self._compute_old_log_prob(batch, metrics=metrics)
-
-        # 6. [OPTIONAL] compute ref_log_prob
-        if self.use_reference_policy:
-            with marked_timer("ref", timing_raw, color="olive"):
-                batch = self._compute_ref_log_prob(batch, metrics=metrics)
-
-        # 7. [OPTIONAL] compute critic values
-        if self.use_critic:
-            with marked_timer("values", timing_raw, color="cyan"):
-                batch = self._compute_values(batch, metrics=metrics)
-
-        # 8. compute advantage and return
-        with marked_timer("adv", timing_raw, color="brown"):
-            batch = self._compute_advantage(batch, metrics=metrics)
-
-        # 9. [OPTIONAL] update critic
-        if self.use_critic:
-            with marked_timer("update_critic", timing_raw, color="pink"):
-                batch = self._update_critic(batch, metrics=metrics)
-
-        # 10. update actor
-        if self.config.trainer.critic_warmup <= self.global_steps:
-            with marked_timer("update_actor", timing_raw, color="red"):
-                batch = self._update_actor(batch, metrics=metrics)
-
-        return batch
-
-
-@ray.remote
-class TaskRunner:
-    def __init__(self) -> None:
-        # role => worker class
-        self.role_worker_mapping = {}
-        # role => resource pool
-        self.mapping = {}
-
-    def add_actor_rollout_worker(self, config):
-        """Add actor rollout worker to mapping."""
-        lora_rank = config.actor_rollout_ref.model.get("lora", {}).get("rank", 0)
-        if lora_rank <= 0:
-            lora_rank = config.actor_rollout_ref.model.get("lora_rank", 0)
-        ref_in_actor = lora_rank > 0 or config.actor_rollout_ref.model.get("lora_adapter_path") is not None
-
-        role = Role.ActorRolloutRef if need_reference_policy(config) and not ref_in_actor else Role.ActorRollout
-        self.role_worker_mapping[role] = ray.remote(ActorRolloutRefWorker)
-        self.mapping[role] = "global_pool"
-
-    def add_critic_worker(self, config):
-        """Add critic worker to mapping."""
-        if need_critic(config):
-            self.role_worker_mapping[Role.Critic] = ray.remote(TrainingWorker)
-            self.mapping[Role.Critic] = "global_pool"
-
-    def init_resource_pool_mgr(self, config):
-        """Initialize resource pool manager."""
-
-        # Global resource pool is used for actor, rollout, critic, ref
-        global_pool_id = "global_pool"
-        resource_pool_spec = {
-            global_pool_id: [config.trainer.n_gpus_per_node] * config.trainer.nnodes,
-        }
-
-        # Add separate resource pool for reward model if enabled
-        if config.reward.reward_model.enable_resource_pool:
-            if config.reward.reward_model.n_gpus_per_node <= 0:
-                raise ValueError("config.reward.reward_model.n_gpus_per_node must be greater than 0")
-            if config.reward.reward_model.nnodes <= 0:
-                raise ValueError("config.reward.reward_model.nnodes must be greater than 0")
-
-            reward_pool = [config.reward.reward_model.n_gpus_per_node] * config.reward.reward_model.nnodes
-            resource_pool_spec["reward_pool"] = reward_pool
-            self.mapping[Role.RewardModel] = "reward_pool"
-        else:
-            config.reward.reward_model.nnodes = config.trainer.nnodes
-            config.reward.reward_model.n_gpus_per_node = config.trainer.n_gpus_per_node
-            self.mapping[Role.RewardModel] = "global_pool"
-
-        distillation_config = config.get("distillation")
-        if is_distillation_enabled(distillation_config):
-            if distillation_config.n_gpus_per_node <= 0:
-                raise ValueError("config.distillation.n_gpus_per_node must be greater than 0")
-            if distillation_config.nnodes <= 0:
-                raise ValueError("config.distillation.nnodes must be greater than 0")
-
-            teacher_pool = [distillation_config.n_gpus_per_node] * distillation_config.nnodes
-            resource_pool_spec["teacher_pool"] = teacher_pool
-            self.mapping[Role.TeacherModel] = "teacher_pool"
-
-        self.resource_pool_manager = ResourcePoolManager(resource_pool_spec=resource_pool_spec, mapping=self.mapping)
-
-    def run(self, config):
-        pprint(OmegaConf.to_container(config, resolve=True))
-        OmegaConf.resolve(config)
-
-        # initialize transfer queue
-        tq.init(config.transfer_queue)
-        trainer = None
-        try:
-            self.add_actor_rollout_worker(config)
-            self.add_critic_worker(config)
-            self.init_resource_pool_mgr(config)
-
-            trainer = PPOTrainer(
-                config=config,
-                role_worker_mapping=self.role_worker_mapping,
-                resource_pool_manager=self.resource_pool_manager,
-            )
-            trainer.init_workers()
-            trainer.fit()
-        finally:
-            tq.close()
-
-
-@hydra.main(config_path="config", config_name="ppo_trainer", version_base=None)
-def main(config):
-    """Main entry point for PPO training with Hydra configuration management.
-
-    Args:
-        config: Hydra configuration dictionary containing training parameters.
-    """
-    # Automatically set `config.trainer.device = npu` when running on Ascend NPU.
-    auto_set_device(config)
-
-    config.transfer_queue.enable = True
-
-    # validate config
-    validate_config(
-        config=config,
-        use_reference_policy=need_reference_policy(config),
-        use_critic=need_critic(config),
-    )
-
-    run_ppo(config, task_runner_class=TaskRunner)
-
-
-if __name__ == "__main__":
-    main()
