+++ verl/trainer/ppo/v1/__init__.py
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from .agent_loop_tq import AgentLoopManagerTQ, AgentLoopWorkerTQ
 from .trainer_base import PPOTrainer, get_trainer_cls, register_trainer
 from .trainer_colocate_async import PPOTrainerColocateAsync
 from .trainer_separate_async import PPOTrainerSeparateAsync
@@ -24,4 +25,6 @@ __all__ = [
     "PPOTrainerSync",
     "PPOTrainerColocateAsync",
     "PPOTrainerSeparateAsync",
+    "AgentLoopWorkerTQ",
+    "AgentLoopManagerTQ",
 ]
